requirements.txt
gspread
google-auth
feedparser
beautifulsoup4
